Feat : UserController -로그인 기능 구현 -회원 가입 기능 구현 -아이디 중복 체크 기능 구현 -유저 정보 변경 기능 구현(닉네임 변경) -비밀 번호 변경 기능 구현 -비밀 번호 찾기 기능 구현 중 -유저 정보 페이지 이동 미구현 -회원 탈퇴 구현 -로그 아웃 미구현 -아이디 찾기 구현 가능성 토론.(유저 정보가 너무 적습니다.)

+++ src/main/java/com/team/twodari/admin/user/AdminUserService.java
@@ -29,7 +29,7 @@ public class AdminUserService {
 
         UserRoleEntity userRoleEntity = adminUserRoleRepository.findById(userSeq)
                 .orElseThrow(() -> new BusinessLogicException(ExceptionCode.USER_ROLE_NOT_FOUND));
-        userRoleEntity.setRoleSeq(UserRoleConfig.UserRole.BLOCK.getLevel()); // 차단
+        userRoleEntity.setRoleSeq((UserRoleConfig.UserRole.BLOCK.getLevel().intValue())); // 차단
 
         return adminUserRoleRepository.save(userRoleEntity);
     }
